# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Movank API
  version: 0.5.0
  description: Productos, ventas, pagos simulados y dashboard por comercio. Importes en centavos de COP.
servers:
  - url: http://localhost:8080
paths:
  /v1/products:
    post:
      operationId: createProduct
      summary: Crea un producto del comercio autenticado
      description: SKU normalizado y único por comercio. No admite merchant_id ni parámetros de query. Cuerpo máximo de 8192 bytes.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductInput'
      responses:
        '201':
          description: Producto creado
          headers:
            Location:
              schema:
                type: string
              description: Ruta del producto creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/SKUConflict'
        '413':
          $ref: '#/components/responses/BodyTooLarge'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
        '503':
          $ref: '#/components/responses/Unavailable'
    get:
      operationId: listProducts
      summary: Lista productos del comercio autenticado
      description: Ordenados por UUID ascendente. Rechaza parámetros desconocidos o repetidos. La paginación no es una instantánea ante altas concurrentes.
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
        - in: query
          name: after
          description: UUID indicado por next_cursor en la página anterior
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Página del catálogo; next_cursor es null al terminar
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [items, next_cursor]
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Product'
                  next_cursor:
                    type: string
                    format: uuid
                    nullable: true
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '503':
          $ref: '#/components/responses/Unavailable'
  /v1/products/{id}:
    get:
      operationId: getProduct
      summary: Consulta un producto del comercio autenticado
      description: Devuelve el mismo 404 si no existe o pertenece a otro comercio. No admite query.
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Producto encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '503':
          $ref: '#/components/responses/Unavailable'
  /v1/me:
    get:
      operationId: currentIdentity
      summary: Consulta la identidad y el comercio del token verificado
      description: El comercio se obtiene de merchant_id en el JWT, nunca del cuerpo, query ni encabezados personalizados.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Identidad autenticada y comercio existente
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [user_id, merchant]
                properties:
                  user_id:
                    type: string
                    format: uuid
                  merchant:
                    type: object
                    additionalProperties: false
                    required: [id, name]
                    properties:
                      id:
                        type: string
                        format: uuid
                      name:
                        type: string
        '401':
          description: Token ausente, inválido o vencido
          headers:
            WWW-Authenticate:
              schema:
                type: string
                enum: [Bearer]
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthError'
        '403':
          description: El comercio del token ya no existe
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthError'
        '503':
          description: No se pudo consultar el comercio
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthError'
  /healthz:
    get:
      operationId: health
      summary: Comprueba que el proceso HTTP responde
      responses:
        '200':
          description: Proceso activo, independientemente de PostgreSQL y Redis
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Healthy'
  /readyz:
    get:
      operationId: readiness
      summary: Comprueba conectividad con PostgreSQL con un plazo de un segundo
      description: No verifica migraciones ni Redis. La respuesta no incluye detalles de conexión.
      responses:
        '200':
          description: PostgreSQL responde
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Healthy'
        '503':
          description: PostgreSQL no responde o se agotó el plazo
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [status]
                properties:
                  status:
                    type: string
                    enum: [unavailable]
  /v1/sales:
    post:
      summary: Crear una venta con idempotencia persistente
      description: >-
        Precios calculados desde el catálogo del comercio autenticado. Máximo 100
        productos distintos y 16384 bytes de JSON. El orden de los productos y
        las mayúsculas de UUID no cambian el pedido. Misma clave y pedido devuelve
        la venta original con 201; distinto pedido devuelve 409. Los errores de
        validación no reservan la clave. Ante timeout o 503, reintentar con la misma clave y pedido.
        No procesa pagos. No se permite merchant_id ni precios en el cuerpo.
      security:
        - bearerAuth: []
      parameters:
        - in: header
          name: Idempotency-Key
          required: true
          description: Clave estable por operación y comercio; sin caducidad.
          schema:
            type: string
            pattern: '^[A-Za-z0-9._:-]{1,128}$'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SaleInput'
      responses:
        '201':
          description: Venta creada o repetición de la venta original
          headers:
            Location:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Sale'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          description: Clave ya utilizada con otro pedido (idempotency_conflict)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductError'
        '413':
          description: Cuerpo mayor a 16384 bytes (body_too_large)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductError'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
        '503':
          $ref: '#/components/responses/Unavailable'
  /v1/sales/{id}:
    get:
      summary: Consultar una venta del comercio autenticado
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Venta con sus precios originales
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Sale'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '503':
          $ref: '#/components/responses/Unavailable'
  /v1/sales/{id}/pay:
    post:
      summary: Pagar una venta con proveedor simulado
      description: >-
        Un intento por venta y comercio, sin requerir otra Idempotency-Key.
        Repetir método y escenario devuelve el estado actual del mismo intento;
        cambiarlos devuelve 409. La referencia se persiste antes de llamar al proveedor.
        TIMEOUT produce UNKNOWN, nunca DECLINED. El worker consulta esa referencia
        y el simulador permite verla como APPROVED después de 3 segundos.
        Si la consulta falla, conserva UNKNOWN y registra cada comprobación.
        Un PENDING interrumpido se reenvía con la misma referencia deduplicada.
        No hay un segundo intento ni reembolsos. Ante 503, consultar la venta o
        repetir exactamente el mismo pago: la confirmación puede haberse persistido.
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [method, scenario]
              properties:
                method:
                  type: string
                  enum: [CARD]
                scenario:
                  type: string
                  enum: [APPROVED, DECLINED, TIMEOUT]
      responses:
        '200':
          description: Resultado definitivo, aprobado o rechazado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Payment'
        '202':
          description: PENDING o UNKNOWN, recuperado por el worker; consultar GET de la venta
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Payment'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Otro método o escenario ya registrado (payment_conflict)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductError'
        '413':
          $ref: '#/components/responses/BodyTooLarge'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
        '503':
          $ref: '#/components/responses/Unavailable'
  /v1/dashboard/today:
    get:
      summary: Agregado de pagos aprobados del día UTC
      description: >-
        Solo cuenta APPROVED por fecha de confirmación. Caché de 30 segundos,
        actualizada por outbox y reconstruible desde PostgreSQL. Puede haber
        un breve retraso respecto al pago. No admite parámetros de comercio.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Dashboard del comercio autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dashboard'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '503':
          $ref: '#/components/responses/Unavailable'
  /v1/dashboard/stream:
    get:
      summary: Actualizaciones SSE del dashboard
      description: >-
        Usar fetch con Authorization Bearer; no se aceptan tokens ni comercios en URL.
        Envía una instantánea inicial y eventos dashboard con el agregado completo,
        id fecha:paid_sales y heartbeat cada 15 segundos. Al reconectar siempre
        envía la instantánea actual; Last-Event-ID no solicita historial.
        El cliente reemplaza el agregado, no suma eventos. Se cierra al vencer
        el JWT, detener el servicio o bloquearse la escritura durante 5 segundos.
        Límite local de 100 conexiones por comercio y 1000 en total.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Flujo SSE del comercio autenticado
          content:
            text/event-stream:
              schema:
                type: string
              example: "id: 2026-09-30:1\nevent: dashboard\ndata: {\"date\":\"2026-09-30\",\"currency\":\"COP\",\"paid_sales\":1,\"total_minor\":20000}\n\n"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          description: Límite de conexiones SSE alcanzado (too_many_streams)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductError'
        '503':
          $ref: '#/components/responses/Unavailable'
components:
  responses:
    BadRequest:
      description: JSON, campos o parámetros inválidos
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProductError'
    Unauthorized:
      description: Token ausente, inválido o vencido
      headers:
        WWW-Authenticate:
          schema:
            type: string
            enum: [Bearer]
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProductError'
    Forbidden:
      description: El comercio autenticado ya no existe
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProductError'
    NotFound:
      description: Recurso no encontrado en el comercio autenticado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProductError'
    SKUConflict:
      description: El SKU normalizado ya existe en este comercio
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProductError'
    BodyTooLarge:
      description: Cuerpo mayor a 8192 bytes
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProductError'
    UnsupportedMediaType:
      description: Se requiere Content-Type application/json
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProductError'
    Unavailable:
      description: No se pudo completar la operación en PostgreSQL
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProductError'
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT HS256 con iss=movank-local, aud=movank-api, exp, iat, nbf, sub (UUID) y merchant_id (UUID).
  schemas:
    Payment:
      type: object
      additionalProperties: false
      required: [id, sale_id, method, scenario, status, amount_minor, checks, last_error, created_at, updated_at, completed_at]
      properties:
        id:
          type: string
          format: uuid
          description: Referencia persistente usada también ante el proveedor
        sale_id:
          type: string
          format: uuid
        method:
          type: string
          enum: [CARD]
        scenario:
          type: string
          enum: [APPROVED, DECLINED, TIMEOUT]
        status:
          type: string
          enum: [PENDING, UNKNOWN, APPROVED, DECLINED]
        amount_minor:
          type: integer
          format: int64
        checks:
          type: integer
          minimum: 0
        last_error:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
          nullable: true
    Dashboard:
      type: object
      additionalProperties: false
      required: [date, currency, paid_sales, total_minor]
      properties:
        date:
          type: string
          format: date
          description: Día UTC
        currency:
          type: string
          enum: [COP]
        paid_sales:
          type: integer
          format: int64
          minimum: 0
        total_minor:
          type: integer
          format: int64
          minimum: 0
    SaleInput:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items:
          type: array
          minItems: 1
          maxItems: 100
          description: Cada producto debe aparecer una sola vez y pertenecer al comercio.
          items:
            type: object
            additionalProperties: false
            required: [product_id, quantity]
            properties:
              product_id:
                type: string
                format: uuid
              quantity:
                type: integer
                minimum: 1
                maximum: 10000
    Sale:
      type: object
      additionalProperties: false
      required: [id, total_minor, currency, created_at, items]
      properties:
        payment:
          type: object
          nullable: true
          description: Solo en GET; null si aún no se ha iniciado un pago. POST conserva la respuesta de creación.
          allOf:
            - $ref: '#/components/schemas/Payment'
        id:
          type: string
          format: uuid
        total_minor:
          type: integer
          format: int64
          minimum: 1
          maximum: 1000000000000
          description: Total en centavos de COP
        currency:
          type: string
          enum: [COP]
        created_at:
          type: string
          format: date-time
        items:
          type: array
          minItems: 1
          maxItems: 100
          items:
            type: object
            additionalProperties: false
            required: [product_id, quantity, product_name, unit_price_minor, subtotal_minor]
            properties:
              product_id:
                type: string
                format: uuid
              quantity:
                type: integer
                minimum: 1
                maximum: 10000
              product_name:
                type: string
              unit_price_minor:
                type: integer
                format: int64
              subtotal_minor:
                type: integer
                format: int64
    ProductInput:
      type: object
      additionalProperties: false
      required: [sku, name, price_minor, currency]
      properties:
        sku:
          type: string
          description: Se recortan espacios y convierte a mayúsculas; resultado de 1 a 64 caracteres A-Z, 0-9, punto, guion o guion bajo, comenzando por letra o número.
          example: CAFE-01
        name:
          type: string
          description: De 1 a 120 caracteres después de recortar espacios exteriores
          example: Café
        price_minor:
          type: integer
          format: int64
          minimum: 1
          maximum: 1000000000000
          description: Centavos de COP; debe enviarse como entero sin decimales
          example: 125050
        currency:
          type: string
          enum: [COP]
    Product:
      type: object
      additionalProperties: false
      required: [id, sku, name, price_minor, currency, created_at]
      properties:
        id:
          type: string
          format: uuid
        sku:
          type: string
          pattern: '^[A-Z0-9][A-Z0-9._-]{0,63}$'
        name:
          type: string
          minLength: 1
          maxLength: 120
        price_minor:
          type: integer
          format: int64
          minimum: 1
          maximum: 1000000000000
        currency:
          type: string
          enum: [COP]
        created_at:
          type: string
          format: date-time
    ProductError:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: string
          enum: [invalid_query, invalid_request, invalid_json, invalid_product, invalid_sale, invalid_payment, payment_conflict, too_many_streams, idempotency_conflict, body_too_large, unsupported_media_type, sku_conflict, not_found, unauthorized, forbidden, unavailable]
    AuthError:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: string
          enum: [unauthorized, forbidden, unavailable]
    Healthy:
      type: object
      additionalProperties: false
      required: [status]
      properties:
        status:
          type: string
          enum: [ok]